refactor: extract call-transport, bypass, IP-channels and reachability renderers
